RDKTV-10842: Observed "Segmentation fault" with BleRcuConsole audio capture
Reason for change:  in BleRcuConsole app if audio stream is stopped
after its already timed-out and terminated in BleRcuDaemon a crash
occurred.  This is fixed by setting the offending pointer to null after
deleting it so as to avoid referencing a deleted object.

I also added spaces to the console prints
Test Procedure:     see ticket
Risks:              Low

Signed-off-by: Gene Gallagher <Eugene_Gallagher@comcast.com>
Change-Id: I692ef887423207b9c910fa79b010281e936898be