LLAMA-5245: [Qorvo Mac] Remote pairing is lost after Llama X3 converter reboot
Reason for change:  When Sky remotes are paired, the advertise name
changes from preprended by a "U" to prepended by a "P".  After a reboot,
the advertise name wasn't matching the contents of config.rdk.json,
therefore preventing the remote from reconnecting properly.  So this
change allows for U or P to prepend the name.
Test Procedure:     using Qorvo MAC, remote is able to be paired and
reconnects successfully with voice working after a reboot.
Risks:              Low

Signed-off-by: Gene Gallagher <Eugene_Gallagher@comcast.com>