LLAMA-9000: IR initiated BLE pairing targets PR1
Reason for change:  IR initiated BLE pairing inserts the 3 digit pair
code to the pairing name so that a specific remote can be targetted.
The 3 digit code was not being used for PR1, which means it could target
a PR1 instead of the desired remote.  This is fixed by adding 3 digit
code to the pairingNameFormat.  Its only used for legacy sky pairing, so
it does not affect PR1 pairing using the MAc hash

Test Procedure:     3 digit code is present for the PR1 when trying to
pair an LC103 using IR initiated pairing.  i.e. in this line there is a
3 digit code in front of PR1. Since the PR1 does not use this pairing
method, the pairing state machine will not try to pair to it.

added pairing regex for supported remote '163 Platco PR1'

Risks:              Low

Signed-off-by: Gene Gallagher <Eugene_Gallagher@comcast.com>
Change-Id: Ibab8967a2685afa45b0680e8bfa06859ad17f9d3